flytter TransportIcons til app/(admin)/components/ og separerer logikk, svg og selve komponenten
