+++ tavla/app/(admin)/components/TransportIcon/index.ts
@@ -1,0 +1,1 @@
+export { default as TransportIcon } from './TransportIcon'
