+++ tavla/app/(admin)/tavler/[id]/rediger/components/Settings/components/TransportPaletteSelect.tsx
@@ -1,6 +1,7 @@
 'use client'
 import { Radio, RadioGroup } from '@entur/form'
 import { Heading4, Paragraph } from '@entur/typography'
+import { TransportIcon } from 'app/(admin)/components/TransportIcon'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import { useEffect, useState } from 'react'
 import type { BoardTheme, TransportPalette } from 'src/types/db-types/boards'
@@ -13,7 +14,6 @@ import {
     generateTransportPalettes,
     getTransportColorDescription,
 } from '../colorPalettes'
-import { TransportIcon } from './TransportIcon'
 
 const busAndTrainModes: {
     mode: TTransportMode
